requirements.txt
flake8==3.7.8
pep8==1.7.1
python-jenkins==1.5.0
PyYAML==5.1.2
texttable==1.6.2
